Create new getCookies selector
- Use cookie whitelist logic from fetchComponents for this selector
- Implement selector in getRouteMeta and getRouteKey
